+++ src/lib/packages/calendar/ui/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @cluue/calendar
 
+## 2.0.7
+
+### Patch Changes
+
+-   Upgrade popover and perf calendar
+-   Updated dependencies
+    -   @cluue/popover@1.0.21
+    -   @cluue/icons@2.0.6
+    -   @cluue/forms@2.0.6
+
 ## 2.0.6
 
 ### Patch Changes