+++ src/lib/packages/popover/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cluue/popover
 
+## 1.0.21
+
+### Patch Changes
+
+-   Upgrade popover and perf calendar
+
 ## 1.0.20
 
 ### Patch Changes